refactor: Compose all components in the Page file.
